Added cage

+++ Rect.cpp
@@ -8,7 +8,7 @@ SDL_Rect Rect::getSDLRect() {
     return SDL_Rect{
         static_cast<int>(x),
         static_cast<int>(y),
-        static_cast<int>(w),
-        static_cast<int>(h)
+        w,
+        h
     };
 }
